Move grid related histroy list to Grids component
